+++ packages/components/components/elvis-accordion/react.d.ts
@@ -1,0 +1,5 @@
+declare module '@elvia/elvis-accordion/react';
+export { default as Accordion } from './dist/react/js/elvia-accordion';
+export type { AccordionProps } from './dist/react/js/elvia-accordion';
+export type * from './dist/react/js/elvia-accordion';
+export type * from './dist/react/js/elvia-accordion.types';