+++ packages/components/components/elvis-dropdown/src/react/elvia-dropdown.tsx
@@ -23,7 +23,7 @@ import {
 import uniqueId from 'lodash.uniqueid';
 import isEqual from 'lodash.isequal';
 import { getColor } from '@elvia/elvis-colors';
-import { ElvisComponentWrapper } from '@elvia/elvis-component-wrapper/src/elvia-component';
+import type { ElvisComponentWrapper } from '@elvia/elvis-component-wrapper';
 import { warnDeprecatedProps } from '@elvia/elvis-toolbox';
 import { config } from './config';
 
@@ -46,10 +46,10 @@ export interface DropdownProps {
    */
   defaultValue?: never;
   value?: DropdownItem | Array<DropdownItem> | undefined;
-  isCompact: boolean;
-  isDisabled: boolean;
-  isMulti: boolean;
-  isSearchable: boolean;
+  isCompact?: boolean;
+  isDisabled?: boolean;
+  isMulti?: boolean;
+  isSearchable?: boolean;
   hasSelectAllOption?: boolean;
   selectAllOption?: Partial<DropdownItem>;
   allOptionsSelectedLabel?: string;
@@ -61,7 +61,7 @@ export interface DropdownProps {
   placeholderIcon?: string;
   valueOnChange?: (selectedOptions: DropdownItem | Array<DropdownItem> | undefined) => void;
   className?: string;
-  inlineStyle?: { [style: string]: CSSProperties };
+  inlineStyle?: CSSProperties;
   webcomponent?: ElvisComponentWrapper;
 }
 
@@ -119,7 +119,7 @@ const Dropdown: React.FC<DropdownProps> = function ({
   const selectId = uniqueId('ewc-dropdown-');
 
   /** Styling functions for react select */
-  const decideControlBorder = (disabled: boolean, error: boolean) => {
+  const decideControlBorder = (disabled?: boolean, error?: boolean) => {
     if (disabled) {
       return `1px solid ${getColor('disabled')}`;
     }
@@ -155,7 +155,7 @@ const Dropdown: React.FC<DropdownProps> = function ({
     return getColor('grey-05');
   };
 
-  const decideSingleValueColor = (isMenuOpen: boolean, searchable: boolean, disabled: boolean) => {
+  const decideSingleValueColor = (isMenuOpen: boolean, searchable: boolean, disabled?: boolean) => {
     if (disabled) {
       return getColor('disabled');
     }
@@ -390,7 +390,7 @@ const Dropdown: React.FC<DropdownProps> = function ({
           {allOptionsHaveIconAttribute ? (
             <Icon
               inlineStyle={{ marginRight: '16px' }}
-              name={(props.data as DropdownItem).icon}
+              name={(props.data as DropdownItem).icon as string}
               size={isCompact ? 'xs' : 'sm'}
               color={optionIsDisabled ? getColor('disabled') : undefined}
             />
@@ -472,12 +472,15 @@ const Dropdown: React.FC<DropdownProps> = function ({
   };
 
   const ElviaSingleValue = (props: SingleValueProps) => {
+    if (menuIsOpen && isSearchable) {
+      return null;
+    }
     return (
       <components.SingleValue {...props}>
         {allOptionsHaveIconAttribute ? (
           <Icon
             inlineStyle={{ marginRight: '16px' }}
-            name={(props.data as DropdownItem).icon}
+            name={(props.data as DropdownItem).icon as string}
             size={isCompact ? 'xs' : 'sm'}
             color={decideSingleValueColor(menuIsOpen, isSearchable, isDisabled)}
           />