+++ packages/elvis-assets-icons/tasks/icons.js
@@ -58,8 +58,10 @@ async function createIconModule() {
 // ADD OR REMOVE ICONS IN icons.config.js
 import { getColor } from '@elvia/elvis-colors';
 `;
-  let iconTypes = `declare type Icon = {
-  getIcon: (color: string) => string;
+  let iconTypes = `import type { ElviaColor } from '@elvia/elvis-colors';
+
+declare type Icon = {
+  getIcon: (color?: ElviaColor | 'inverted') => string;
 };`;
   let iconNameType = `
 export declare type IconName =`;
@@ -93,9 +95,9 @@ export const ${createCamelCase(icon.name)} = {
       icon = icon.replace(/fillGreen/g, "fill='#29D305'");
       return icon;
     } else if (!color.startsWith('#')) {
-      return icon.replace(/fill="([^"]*)"/g, 'fill="' + getColor(color) + '"');
+      return icon.replace(/fill="#000"/g, 'fill="' + getColor(color) + '"');
     }
-    return icon.replace(/fill="([^"]*)"/g, 'fill="' + color + '"');
+    return icon.replace(/fill="#000"/g, 'fill="' + color + '"');
   },
 };`;
     iconTypes =
@@ -162,9 +164,9 @@ async function createCommonJSIconModule() {
             icon = icon.replace(/fillGreen/g, "fill='#29D305'");
             return icon;
           } else if (!color.startsWith('#')) {
-            return icon.replace(/fill="([^"]*)"/g, 'fill="' + getColor(color) + '"');
+            return icon.replace(/fill="#000"/g, 'fill="' + getColor(color) + '"');
           }
-          return icon.replace(/fill="([^"]*)"/g, 'fill="' + color + '"');
+          return icon.replace(/fill="#000"/g, 'fill="' + color + '"');
       }
     }`;
     iconTypes =