+++ packages/elvis-assets-icons/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Elvis-assets-icon
 
+## 2.4.2 (16.08.22)
+
+### Patch
+
+- Fixed the type declaration for the `getIcon`-function.
+
+## 2.4.1 (15.08.22)
+
+### Bug fix
+
+- Fixed a bug that caused the color property to not work properly for filled icons.
+
 ## 2.4.0 (05.08.22)
 
 ### New feature